LOADS of changes
Ok so there were a lot of things that I went over and changed.
The first thing is the check to see if the user inputted nothing. So at
first I just used the "if not" statement, but this only meant if they
just pressed enter then tell them to enter something. So what if they
only pressed spacebar and then pressed enter? To fix this I used the
string module so I could check if any letter in the alphabet was in the
string and if it wasn't, tell them to enter something. This makes a lot
more logical sense :)

The second thing I did was go through my task 3. I realised that after
all these changes I've been making to task 1 and 2, I haven't been
updating the code that's relevant in task 3. So I just did as much of
that as I could but there could be a little more to do. So yeah, a few
changes and a massive task 3 update :+1: